requirements.txt
numpy==1.21.5
opencv_python==4.6.0.66
pandas==1.4.2
Pillow==9.2.0
streamlit==1.11.0
